chore: reorganize file structure by FSD
