+++ personalization-sdk/src/main/kotlin/com/personalization/inAppNotification/view/component/image/ImageView.kt
@@ -11,7 +11,7 @@ import com.bumptech.glide.request.RequestListener
 import com.bumptech.glide.request.RequestOptions
 import com.bumptech.glide.request.target.Target
 
-class AlertImageView @JvmOverloads constructor(
+class ImageView @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null,
     defStyleAttr: Int = 0
