+++ personalization-sdk/src/main/res/layout/alert_button_container.xml
@@ -15,7 +15,7 @@
         android:layout_weight="1"
         app:cardCornerRadius="@dimen/small_corner_radius">
 
-        <com.personalization.inAppNotification.view.component.button.AlertButton
+        <com.personalization.inAppNotification.view.component.button.Button
             android:id="@+id/buttonDecline"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
@@ -34,7 +34,7 @@
         android:layout_weight="1"
         app:cardCornerRadius="@dimen/small_corner_radius">
 
-        <com.personalization.inAppNotification.view.component.button.AlertButton
+        <com.personalization.inAppNotification.view.component.button.Button
             android:id="@+id/buttonAccept"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
