+++ personalization-sdk/src/main/res/layout/alert_text_container.xml
@@ -4,14 +4,14 @@
     android:orientation="vertical"
     android:layout_height="match_parent">
 
-    <com.personalization.inAppNotification.view.component.text.AlertText
+    <com.personalization.inAppNotification.view.component.text.Text
         android:id="@+id/title"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:text="@string/alert_dialog_title"
         android:textStyle="bold" />
 
-    <com.personalization.inAppNotification.view.component.text.AlertText
+    <com.personalization.inAppNotification.view.component.text.Text
         android:id="@+id/message"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
